backend/api-gateway-fastapi/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
prometheus-client==0.20.0
python-dotenv==1.0.1
python-jose==3.3.0
